Differentiate post id from repost id
