+++ src/pages/MyLikes.js
@@ -40,7 +40,7 @@ function MyLikes() {
         ) : (
           posts.map((post) => (
             <Post
-              key={post.id}
+              key={post.repostId ? post.repostId : post.id}
               username={post.user.username}
               text={post.text}
               link={post.link}
@@ -51,11 +51,12 @@ function MyLikes() {
               likes={post.likes}
               userId={post.user.id}
               id={post.id}
+              repostId={post.repostId ? post.repostId : false}
               setNewPosts={setNewPosts}
               newPosts={newPosts}
               repostCount={post.repostCount}
               repostedByUser={post.repostedBy?.username}
-              repostedById={post.repostedBy?.id}
+              repostedUserId={post.repostedBy?.id}
             />
           ))
         )}
