+++ src/pages/MyPosts.js
@@ -38,7 +38,7 @@ function MyPosts() {
         ) : (
           posts.map((post) => (
             <Post
-              key={post.id}
+              key={post.repostId ? post.repostId : post.id}
               username={post.user.username}
               text={post.text}
               link={post.link}
@@ -49,11 +49,12 @@ function MyPosts() {
               likes={post.likes}
               userId={post.user.id}
               id={post.id}
+              repostId={post.repostId ? post.repostId : false}
               setNewPosts={setNewPosts}
               newPosts={newPosts}
               repostCount={post.repostCount}
               repostedByUser={post.repostedBy?.username}
-              repostedById={post.repostedBy?.id}
+              repostedUserId={post.repostedBy?.id}
             />
           ))
         )}
